Check yum repo syntax
Rename '_use_unvetted_yum_repos' to '_check_yum_repos' so we can perform
more sanity checks in the same parse.

Previously we were only checking for unvetted repo enabled on the system.
We are now in addition to that also checking for invalid syntax.

'\$' does not behave the same between RedHat 7 and 8.
RedHat 7 interpolates the variable whereas RedHat 8 does not.