Avoid duplicate output on errors:
* `run_service_and_notify()` calls `_run_service()`
* If `_run_service()` fails, `run_service_and_notify()` will call `LOGDIE()`
  with whatever error was provided - this logs an error and throws an exception
* `_run_service()` and all the subroutines it runs (such as `run_stage_3()`
  that handles LEAPP) themselves also call `LOGDIE()` if there's an error,
  which logs an error and throws an exception

This means that a `LOGDIE()` is logged and an exception is thrown with an error
logged to the terminal/log file. It is then caught in the caller subroutine and
`LOGDIE()` is called again, causing the same logging message to be logged again.

I tried verifying that every `LOGDIE()` I changed is only called within a stage
so it will be caught by `run_service_and_notify()` and be logged and the process
closed properly.

Resolves GH #36.