Handle log output with log4perl
Originally, logging to a file was done by piping screen output to
`tee(1)`. This prevented some output from being logged, as only the
systemd service did this. Instead, use a native log4perl file appender
to log to the file.

This eliminates colored output, since there is no direct way to save
color data to the file anymore. Restoring this feature will be explored
in a future issue.

Resolves #71 and #83.