Fix reporting of final MySQL upgrade status
Nested definition of `$out` in `reinstall_mysql_packages` is causing
that function to report less useful information in the case of a MySQL
upgrade failure. Report the more useful final status of the upgrade,
rather than the initial return value of the API call which kicked off
the process.

Fixes #75.