Block if BLS is not disabled in /etc/default/grub
By default, the upgrade process has been converting normal GRUB2 boot
loader entries into files compatible with a subset of the Boot Loader
Specification. (A Fedora wiki entry
<https://fedoraproject.org/wiki/Changes/BootLoaderSpecByDefault>
describes the details.) However, server providers may boot into a
version of GRUB2 which does not understand BLS or does not properly
parse these files. If this happens, the upgrade process will cause the
system to no longer boot without manual intervention.

BLS is (currently) optional for AlmaLinux 8, and it is possible to
suppress this conversion by explicitly setting `GRUB_ENABLE_BLSCFG` to
`false` in `/etc/default/grub` prior to the upgrade. Therefore, block
upgrades if this is not set, but also provide a command line flag to
bypass this workaround. Also, explicitly mention the URL of issue #101
in the blocker, so that more information can be gathered.

Fixes #101.