Re-work the BLS blocker
QA reported that #112 was unable to handle double-quoted values, which
it should, because `/etc/default/grub` and similar files are able to be
any valid shell variable assignments, including non-constant expressions.

Since there is no better parser of shell scripts than the shell, use
that to extract the value from the file. Drop UID to `nobody` and run
`bash` in its restricted mode to try to prevent unintended side effects.

Since direct I/O isn't used anymore, switch the tests of the blocker
itself to use Test::MockFile, and write a test for the shell-based
parsing mechanism. This requires rewriting a few of the previous tests.