Work around external GRUB2 issue
It turns out that we cannot instruct users to disable BLS, because
AlmaLinux 8 and other RHEL8-derived distros removed support for
modifying boot entries written in terms of native GRUB2 configuration.
Specifically, the `grubby` utility was dropped in place of a wrapper
which only understands BLS. (The rest of the infrastructure needed to
use native bootloader entries appears to still be in place, but bringing
back the classic `grubby` is not something we want to maintain.)

More careful reading of a thread involving the provider in question
offering a fix for the same issue upon upgrade from Fedora 29 to 30
and the way BLS support is implemented in GRUB2 suggested that the
provider is using a Debian-style version of GRUB2 which requires
`/boot/grub` as the GRUB2 prefix, while RHEL-derived distros use
`/boot/grub2` as the prefix. To work around this, all of their RHEL-like
distro images were being populated with a symlink from
`/boot/grub/grub.cfg` to `/boot/grub2/grub.cfg`. The issue here is that
BLS support is implemented as a GRUB2 module, and symlinking only the
configuration file is insufficient to allow their version of GRUB2 to
find the module. Therefore, `/boot/grub` must be symlinked to
`/boot/grub2`, so that the module can be found and loaded.

Detect when the old workaround is present, convert it to the new
workaround, and copy non-conflicting files from the previous
`/boot/grub` directory into `/boot/grub2` as a courtesy. Warn but
continue if something goes wrong with the copy. Block if it is unclear
whether any workaround for GRUB2 is present or absent.

Fixes #117.