Implement self-update mechanism
Allow users to update to the latest version of the script using the
`--update` flag.

Due to the need for an easy way to identify whether the existing version
of the script on disk has been superseded by a newer one, an explicit
version number is being introduced. Incrementing this is to be performed
by Makefile recipe `release`. Both the `elevate-cpanel` script and the
`version` file will be changed to have the same value.