Sanity check before/after leapp process
Add 'FAQ' and 'SumUp' sections to the website.
Enforce sanity check for stage 2,3 and 4.

If the leapp process failed we end up in Stage 4
without being on a AlmaLinux 8 yet... thus reset
the stage to 3 to avoid being locked.

Reversely we cannot run stage 2 or 3 while being
on major version 8.