Exempt EA4 OpenSSL devel packages from blocker
The EasyApache packages for the development files belonging to the
OpenSSL builds used on CentOS 7 do not depend on the main packages being
present. As such, it is possible for the `ea-openssl-devel` package to
have become an orphan when `ea-openssl` was replaced with
`ea-openssl11`.

Since the absence of these development packages do not cause problems
for cPanel itself, allow the script to ignore the removal of these
packages, at least for the purpose of the EA4 compatibility blocker.

Fixes #119.