Check for warning file from server provider
If the file `/var/cpanel/elevate-noc-recommendations` exists, present
the user with a generic warning asking them to check with their server
provider. If the file has contents, display those instead of the generic
warning. Prompt for an acknowledgement that the user has done what has
been requested, and bail out if this is not answered in the positive.

Implements #198.