Block if the running and boot kernels differ
The leapp framework requires the ability to instruct the boot loader to
load the OS with a custom initrd. If this is not possible, the upgrade
will not work. However, since we cannot directly determine whether we
can affect the boot loader without rebooting[*], use kernel versions as a
proxy.

[*] It is possible to guarantee agency over the boot loader by adding a
dummy kernel parameter, which could be detected after reboot with
`/proc/cmdline`; however, this method was considered to be overkill.