Fix #287
Since this is a bug in cPanel v110 or earlier, there's really no
reasonable expectation to fix this anywhere but "in the mix" so just
enable MySQL before performing the upgrade so that it does not fail.
If it was disabled previously, re-disable it upon upgrade completion,
however that winds up falling out.