Vet the digitialocean-agent repo
Add `digitalocean-agent` to the list of vetted repos. The exact same
package is used on all supported RHEL-like distros, which includes both
7 and 8, so allowing it should not interfere with the upgrade leapp
performs.

Additionally, the re-installation of the packages is not necessary. Both
are Go programs with no dependencies except for glibc. The services
start after the upgrade, and the repositories appear to remain valid.