Block when cP accts have DBs on distro PgSQL
Modify the Databases blocker to add a condition which checks for the
following three conditions:

1. a `postgresql-server` package is installed;
2. the file `/var/cpanel/acknowledge_postgresql_for_elevate` does not
   exist;
3. one or more cPanel users have a PostgreSQL database mapped to them.

If all three are true, issue a blocker. In the blocker message, indicate
which cPanel accounts have relevant databases, point the admin to
resources which explain how to update the PostgreSQL data directory, and
instruct the admin on how to acknowledge the blocking warning using the
touch file. Additionally, modify the "Blockers" documentation page with
the same resources as well as a potential way to use distro-provided
tools to perform the update of the data directory and resolve the issue.