Teach should_run_leapp to check for the CLI option
Previously, the check used by blockers to determine if the script will
run leapp was incompatible with the check option as it only read a file
that was written just prior to running stage 1.  This updates the check
to see if we should run leapp to also check to see if the --no-leapp
option was passed via CLI in order to make the sub compatible with the
check option.

Fixes #332