Add blocker for start mode if backup or upcp is currently executing
Case RE-71: elevate-cpanel adds a motd notice that elevate is in
progress and warns the sysadmin not make any changes while it is
executing.  It also expects to reboot the system multiple times.  As
such, we need to check to ensure that scripts/upcp and bin/backup are
not executing before starting the elevate process.  scripts/upcp is
especially problematic if it is running simultaneously as the reboot
could leave the system in a very difficult to repair broken state.

Changelog: Add blocker for start mode if bin/backup or scripts/upcp is
 currently executing