Convert ssh blocker to an autofix
Case RE-112: For sshd, the default value for PermitRootLogin
changes from "yes" to "no" when upgrading from CentOS 7 to a
CentOS 8 variant. We used to block if PermitRootLogin was not
explicitly set; but, instead it is more desirable to simply
emit a warning that we are going to set it to "yes" and go ahead
and set it so the behavior stays the same post upgrade. I've
altered the SSH blocker module to emit the warning and not block.
I've also moved the SSH blocker execution to happen earlier so
the warning is more visible.  And, I've added a component that
will provide the autofix for the sshd configuration file.

Changelog: Set PermitRootLogin for sshd to 'yes' if not explicitly
  set in the configuration file so that the behavior does not
  change after the upgrade.