Teach Elevate::OS:instance() to emit a user friendly error
Case RE-18:  This is a request from the review.  Instead of relying on
Elevate::OS::is_supported() to be the first call into Elevate::OS, this
change makes it so that Elevate::OS::instance() will always emit a user
friendly error when it dies due to the OS not being supported.

Changelog: