Convert Elevate::OS to use constants rather than accessors
Case RE-18:  This is a request from the review.  Since most of the data
provided by Elevate::OS is constant, we should use constants rather than
accessors to provide this data.  There is less code involved this way
and it makes the code cleaner/easier to read over time.

Changelog: