Add blocker for invalid CloudLinux licenses
Case RE-173:  CloudLinux's fork of ELevate verifies that the CloudLinux
licese is valid and inhibits the elevation process from starting if it
is not.  Due to this, we need to block if the CloudLinux license is not
valid in order to avoid leaving the system in a broken state.

Additionally, I fixed a bug where the argument would not be numeric in
exit when blockers were set to abort on the first blocker, and a bug
where the blocker was not added before dying in this same situation
which would result in the elevate-blockers file essentially being null.

Changelog: Add blocker for invalid CloudLinux licenses