Report back the status of rpmup if yum is out of date
Case RE-89:  If yum is out of date, we need to know the status of rpmup
as this data will help us to determine if rpmup is failing to update the
packages or if it is disabled resulting in the packages being out of
date.

Additionally, in order to better see the data, it should always be
returned as an aref.  For a developer on the script itself, it feels
more like it should be an href, but from the standpoint of the analytics
software, it should be an aref since each server reporting back data is
a single element in the array from the standpoint of that software, so
we essentially send back an aref with one element.

Changelog: