Have script update yum during start mode
Case RE-172: Previously, we would block if yum was not up to date.  This
change makes it so that we attempt to update yum for the customer if the
script is in start mode and report it back as a blocker if yum is unable
to update.

Changelog: Have script update yum during start mode