Skip Database components if DB software is provided by CL
Case RE-153: CL leapp supports updating CL versions of
MySQL/MariaDB/Percona, so we do not need to remove the database packages
provided by CL when we detect that CL MySQL is in use as leapp will
handle this.  However, MySQL Governor does leave the cPanel/upstream
repo files in place as well as the package that provides the repo files
installed when it replaces cPanel/upstream MySQL with CL MySQL.  This
leads to an inhibitor in the leapp preupgrade blocker check that we
perform.  This change adds this inhibitor to the ignore list and adds a
minimal pre_leapp component to remove the packages that provide the repo
files when CL MySQL is detected.

Changelog: