Add support for CL MySQL
Case RE-153: This change detects if CL MySQL or cPanel MySQL is in use.
If cPanel MySQL is in use, then use the previous logic to handle MySQL.
If CL MySQL is in use, then block if MySQL is below version 5.5.
Otherwise, allow the elevate process to continue, and allow leapp to
handle updating MySQL for CL8 as leapp should handle it.

Changelog: Add support for CL MySQL