Convert systems using Sectigo to use Let's Encrypt
Case RE-54:

When CentOS 7 is elevated to AlmaLinux 8, the "nobody"
user will retain its uid/gid of 99 where as on AlmaLinux 8
the expected uid/gid for "nobody" is 65534. This will
cause issues with the Sectigo AutoSSL provider.
If a system is setup to use Sectigo as the AutoSSL provider,
emit a warning that the system will be switched to
Lets Encrypt. Run the appropriate autofixer for the conversion
during the upgrade

Changelog: Change to Let's Encrypt if Sectigo is the AutoSSL provider.