Also block on the 'leapp preupgrade' ERRORS
Case RE-223:

Normally, during a leapp upgrade or preupgrade, leapp
reports any issue when inhibits an upgrade in the
leapp-report.json file and flags it as in "inhibitor".
Sometimes, however, leapp encounters a fatal error
which is not flagged as such in the report file.
Fatal errors will show up in the stdout in a special
"ERRORS" block. Also, if leapp encounters either an
inhibitor or any other fatal error, the exit code will
be non-zero.
So, if we have a non-zero error code, and an error
block reported in leapp's stdout, we should report
this as a blocker.