Have bail_out_on_inappropriate_distro() clear Elevate::OS cache
Case RE-249: Elevate did not detect the correct OS on a server that had
been elevated and still had the staging file in place.  This was due to
the cached OS value still being present in the staging file.  Due to
this, elevate did not bail out as expected on an OS that was not
eligible to be elevated.

Changelog: Ensure proper OS detection on servers that have been
 previously elevated