Report potential failures of LEAPP.
Case RE-224: This commit adds in some extra checks to help determine
if the LEAPP process failed or not. It also introduces a new potential
touch file at /var/cpanel/elevate_leap_fail_continue that can be
used to continue even if the new checks detect a possible error.

Changelog: Added in more checks to report potential failures of
 LEAPP.