Remove unsupported software section from blockers documentation
Case RE-78:  This code in this case converts the CCS blocker to a
component.  This change updates the documentation to remove an
unsupported software section from the blockers documentation page since
the only thing that the unsupported software section contains is that
CCS is unsupported.  Since CCS is now supported, we might as well remove
the entire section from the page.

Changelog: