Ensure all system calls are executable absolute paths
Case RE-260: C::S::O has a bug in 110 where it does not detect missing
scripts (see CPANEL-40343).  The change in CPANEL-40343 was deemed
medium/high risk, and was rejected as a backport candidate for 110.  As
such, we need to work around this issue.  Since elevate-cpanel can only
be executed as root, it was determined that all commands that it
executes should be executabl absolute paths.  This change enforces that.

Changelog:  Ensure all system calls are executable absolute paths