Update documentation to reflect that CloudLinux upgrades are supported
Case RE-144:  In case RE-18, we added support for CloudLinux 7 to
CloudLinux 8 upgrades.  At the time, this was deemed experimental, so we
held off on updating our documentation to reflect this.  However, in
version 33 of this script, the experimental tag for CloudLinux was
removed.  As such, we now need to update the documentation to reflect
that CloudLinux is supported.

Changelog: