Do not block on dropped PHP packages if no domains use them
Case RE-313:  Currently, the EA4 blocker will block if there are any
packages that will be dropped (not provided on the newer OS) during the
elevation.  This change teaches the blocker to check to see if any
domains are using the installed version of PHP.  If there are no domains
using the version of PHP that will be dropped, then we no longer block
for that version of PHP and allow the elevation to continue if that was
the only thing blocking it.

Changelog: