Teach EA4 blocker about hardened PHP provided by Imunify 360
Case RE-313: Imunify 360 is a paid product that provides hardened PHP
versions that are available on newer OS versions.  Currently, the EA4
blocker would block on older PHP versions such as ea-php54 since they
are EOL and not provided on AlmaLinux 8 by cPanel.  However, Imunify 360
provides these versions of PHP on AlmaLlinux 8 as well as 7 so elevate
should not block if the hardened PHP version is provided by Imunify 360.
This change updates the EA4 blocker to no longer block on hardened PHP
versions that are provided by Imunify 360.

Changelog: