Update blocker documentation to reflect EA4 blocker changes
Case RE-313:  The EA4 blocker now detects if dropped PHP versions are in
use and allows the elevation to proceed if they are not in use.  It also
detects if the dropped PHP versions are provided by Imunify 360's
hardened PHP repo as they are not dropped in this specific use case.
This change updates are blocker docs to reflect this.

Changelog: