Have script block if mount -a fails to exit cleanly
Case RE-134: Since the script reboots the server and relies on the file
system to be consistent between reboots, we need to ensure that mount -a
exits cleanly before starting the elevate process.  This change adds a
blocker to the script is mount -a fails to exit cleanly.

Changelog: Have script block if mount -a fails to exit cleanly