Tidy Grub2 blocker
Upstream github tidy and sandbox tidy do not agree on what to do to make
this file tidy.  As such, I just put a blank line in between the lines
that upstream github and the sandbox could not agree on so that they
should both consider the file tidy.

Changelog: