Restore PHP vhost versions post elevate
Case RE-358:  If the version of PHP being installed is provided by
Imunify 360 post elevate, it is possible that any domains using it will
be reset to use the inherited version of PHP instead.  Since this can
break the website for those domains, we need to ensure that they are set
to the correct version of PHP after elevate completes.

Changelog: