Teach EA4 blocker about ea-profiles-cloudlinux provided by Imunify 360
Case RE-362: When Imunify 360 is installed, it is possible to install
the ea-profiles-cloudlinux package which is provided by Imunify 360 and
available on C7 and A8.  Previously, the EA4 blocker would consider
ea-profiles-cloudlinux a dropped package and block based on it being
installed.  This change teaches the blocker to check to see if this
package is provided by Imunify 360 and allows it the elevation to
continue if it is.  It also adds this package to the list of packages
provided by Imunify 360 that need to be reinstalled after the elevation
completes if it is installed and provided by Imunify 360.

Changelog: