Remove blocker for when /usr is a separate private mount point
This change reverts 'c3a46ab218066f85d4753ea4af25a8279a08bce4' which was
added via RE-306.  Since CloudLinux has an upstream fix for this issue,
it is no longer necessary for us to block in this scenario.

Changelog: Remove blocker for when '/usr' is a separate private mount
 point