Add blocker for an invalid Imunify license
Case RE-373: If a server has an invalid Imunify license, then the
Imunify component can fail in expected ways.  This can further lead to
the elevate script failing in stage 4 due to the Imunify post_leapp
component failing and the EA4 component relying on the Imunify component
to succeed.

Changelog: Add blocker for an invalid Imunify license