Convert multiple NICs blocker to a component
Case RE-337:  Previously, we blocked if we found that there were
multiple NICs with names conflicting with the kernel naming space.  This
change makes it so that we give a warning about this when the script is
executed in check mode, and we rename the NICs to a different naming
scheme if the script is executed in start mode.  Of note, we do
explicitely ask the user for permission to make this change on their
behalf before performing the change since there is an inherint risk
involved of networking not coming back up when the server reboots.

Changelog: Convert multiple NICs blocker to component