Update generate_changelog to no longer rely on build-tools
Case RE-447:  It was determined that we no longer want to rely on
build-tools/generate_changelog in order to generate the changelog for a
new version of elevate.  This change makes it so that elevate's
generate_changelog script parses the git logs directly rather than
relying on wrapping 'build-tools/generate_changelog' in order to perform
this task.

Changelog: