Do not allow check to run in the middle of an elevation
Case RE-505:  Previously, a user or security advisor could execute
--check in the middle of a failed elevation.  It does not make sense for
this to be allowed since the elevation attempt has already made it past
the blocker stage.  At this point, it is far more appropriate for the
user to execute --continue in order to complete the elevation.

Of note, one of the side affects of this is that --check will clear the
Elevate::OS cache which will cause failures if the elevation attempt has
already made it to stage 4 or 5 and the OS has been upgraded, but cPanel
services and friends have not been restored.

Changelog:  Do not allow check to run in the middle of an elevation