Add jetapps-* to the list of vetted repos
Case RE-542: In RE-420, we updated the Repository blocker to check for
packages installed from repositories that are no longer enabled.  This
resulted in jetapps support being broken since the jetapps repositories
were not in the list of vetted repositories.  This change reinstates
jetapps support.

Changelog:  Add jetapps-* to the list of vetted repos