Have Imunify 360 check ensure ea-cpanel-tools is provided by CL
Case RE-535: A server with Imunify 360 installed could report that it
had the hardened PHP feature, but either have the repo disabled or not
have updated the ea-cpanel-tools package to be provided by CloudLinux.
This would result in the EA4 blocker check failing because it would
attempt give the EA4 backup script
(/usr/local/bin/ea_current_to_profile) the 'CloudLinux_8' target which
the cPanel version of ea-cpanel-tools does not provide.  This would
result in the command failing due to an invalid target.  This change
makes it so that we verify that ea-cpanel-tools is being provided by
CloudLinux before assuming that Imunify 360 provides hardened PHP.

Fixes #466

Changelog: Have check that Imunify 360 provides hardened PHP ensure that
 the 'ea-cpanel-tools' package is provided by the CloudLinux repository