Add blocker for duplicate repo IDs
Case RE-452: leapp has an inhibitor in place to check for any
repositories that duplicate repo IDs.  This can lead to a situation
where --check recommends that a server is ready for elevation, and
--start blocks since it 'leapp preupgrade' is only executed during
--start.  This change adds a blocker to perform the same check the leapp
performs so that we can block and advise the customer how to remove the
blocker during --check.  Additionally, one of the things that is
triggering this issue on customer server appears to be repo files that
are put in place by the product during MariaDB upgrades.  As such, this
change also adds in an autofix if the duplicate repo ID matches
MariaDBXX since it was likely put in place by the product and will be
removed as part of the pre_leapp component for the databases.

Chagnelog:  Add blocker for duplicate repo IDs