Do not cleanup database packages if the database is provided by CL
Case RE-544: When MySQL governor takes over the MySQL install, it can
leave a mix of packages from the MariaDB/MySQL upstream and
MySQL/MariaDB CloudLinux repos in place.  After discussing this with
CloudLinux's developers, it was determined that CloudLinux's version of
leapp is able to handle this situation and that elevate should NOT
remove these packages as it result in a situation where the database is
NOT installed.  This change makes it so that elevate does NOT remove any
database packages if it determines that the database server is provided
by CloudLinux.

Changelog: Do not cleanup database packages if the database is provided
 by CloudLinux