Do not log output of Imunify license check
Case RE-404: The output of the Imunify license check could be really
noisy in the output while in check mode making it more difficult to pick
out if a blocker has been found.  This change makes it so that the
output of this command is no longer printed to STDOUT or logged.

Changelog: Do not log output of Imunify license check