Restore my.cnf after installing the database server during post_leapp
Case RE-543: The contents of my.cnf have the potential to cause the
restoration of the database server to fail.  As such, we need to restore
the original my.cnf from pre_leapp after restoring the database server
in post_leapp.  Once the custom my.cnf is restored, we should attempt to
restart the database service to make sure that it starts with the
original my.cnf from before elevate began in place.  If it does, then we
can move on.  If it does not, then we need to put the default my.cnf
that is generated by the installation process of the database restore
back in place and make sure that the database service starts with it in
place.  This change makes it so that it is more likely for the database
service to be installed and running after post_leapp component for it
has completed.

Fixes #469

Changelog: Restore my.cnf after installing the database server during
 post_leapp