Add 'check_detected_devices_and_drivers' to list of inhibitors to ignore
Case RE-614: Upstream leapp version 0.19.0 is currently in testing but
pending being released to stable sometime this month.  One of the
changes in this version is the name of the inhibitor that checks the
loaded kernel modules.  It changes from 'check_installed_devel_kernels'
in version 0.16.0 to 'check_detected_devices_and_drivers' in version
0.19.0.  In order to ensure elevate-cpanel supports both versions of
upstream leapp, we simply add 'check_detected_devices_and_drivers' to
the list of inhibitors to ignore so that it is ignored under both names.

Changelog: Add 'check_detected_devices_and_drivers' to list of
 inhibitors to ignore in the leapp preupgrade check to prepare for
 upstream leapp version 0.19.0