Restore EA4 config files based on package prefix
Case RE-499: The restore logic for EA4 config files was broken when the
packages were provided by either CloudLinux's EA4 repo or Imunify's
hardened PHP repo.  This was because the restore logic was gathering the
EA4 packages based on the repoid that they came from, but the regex only
took cPanel's version of EA4 into account.  This change makes it so that
we gather these packages based on the package prefix of 'ea-' so that
the config files for the installed packages are always gathered
regardless of which repo provides them.

Fixes #486

Changelog: Fix restore logic for config files of packages provided by
 EA4