Move logic to gather and store php usage to a component
Case RE-611: With the changes to the logic to support Imunify 360 on
CentOS 7, it was found that sites using older hardened PHP versions no
longer had the correct version of PHP set after elevate completed.  This
was because the changes to the Imunify 360 logic made it so that the
code path that gathered and stored this information was no longer
reachable.  As such, this logic has been refactored to only check for
PHP versions that would otherwise be dropped after the upgrade to
AlmaLinux 8 that are not in use and the logic to gather and store the
current PHP usage has been moved to the EA4 component so that it always
gets executed and thus the logic to set the PHP version after the
elevation will always get executed now too.

Changelog: Move logic to gather and store php usage to a component