Add a whitelist for some legacy packages without a repo
Case RE-603: In RE-420, we started blocking on packages that do not have
an associated repo.  This led to a significant spike in the number of
servers that were eligible to be elevated out of the box.  Upon looking
at the data returned from the blocker added via RE-420, we found that
there were several packages that we could whitelist for this blocker
which would reduce the number of servers that were blocked out of the
box.  This change creates the whitelist and ignores the packages that
were added in it.

Changelog: Add a whitelist for some legacy pacakges without a repo