Ensure Imunify upgrade is not executing before uninstalling it
Case RE-862: The Imunify products have the ability to auto update
themselves which can trigger a lock file being in place that will
prevent the Imunify installer(s) from executing.  This change makes it
so that we check to see if the lock file exists and wait for the PID to
complete before attempting to uninstall Imunify in the
pre_distro_upgrade stage.

Changelog: Ensure Imunify upgrade is not executing before attempting to
 uninstall it