Use mv and cp from File::Copy instead of copy and move
Case RE-723:  As it turns out, File::Copy provides copy and cp that
behave slightly different.  The copy and move subs use the umask
provided by the running processes shell.  The cp and mv try to preserve
permissions for the file that is being copied or moved.  As such, we
should prefer cp and mv over copy and move from File::Copy.

Changelog: Use mv and cp from File::Copy instead of copy and move