Move package manager logic behind abstract class
Case RE-252: This change refactors all the yum/dnf/rpm logic so that it
is all behind Elevate::PkgMgr making it so that callers just need to
execute something such as Elevate::PkgMgr::install() to install
packages.  This will allow us to drop in Elevate::PkgMgr::APT when we
add support for Ubuntu instead of adding a bunch of if/then/else logic
all over the code base.  It should also make it easier to debug any
package manager issues going forward to since they are all consolidated
to one place in the code base.

Changelog: Refactor package manager logic